requirements.txt
# Flask extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.2

